my-nutri-journey-main/backend/requirements.txt
fastapi
uvicorn
langchain-openai
openai
python-dotenv
bcrypt
pandas
diskcache
python-multipart
supabase